Move files to more specific folder and change namespace name
