+++ src/DevSpector.Application/DevicesManagement/ILocationManager.cs
