+++ src/DevSpector.Application/DevicesManagement/LocationManager.cs
